refactor(plugins): share setup presentation URL policy
